refactor:변수 이름 수정 및 쓰지않는 props 정리
