--- src/recoil/productListState.ts
@@ -1,9 +1,0 @@
-// ** 통신시 구현 전역데이터 관리 로직 **
-// import { atom } from 'recoil';
-
-// import { Product } from '../components/MainPage/ListItemComponent/AllProductList';
-
-// export const productListState = atom<Product[]>({
-//   key: 'productListState',
-//   default: [],
-// });
